+++ src/modules/canvas/components/index.ts
@@ -2,3 +2,5 @@ export { GraphCanvas } from './GraphCanvas';
 export { EntityNode, type EntityNodeData } from './EntityNode';
 export { ConnectionEdge, type ConnectionEdgeData } from './ConnectionEdge';
 export { ConnectionInspector } from './ConnectionInspector';
+export { ViolationBadge } from './ViolationBadge';
+export { ViolationTooltip } from './ViolationTooltip';