+++ src/modules/canvas/hooks/index.ts
@@ -4,3 +4,5 @@ export { useConnectionsAsEdges } from './useConnectionsAsEdges';
 export { useConnectionHandlers } from './useConnectionHandlers';
 export { useSelectedConnection } from './useSelectedConnection';
 export { useSuggestedEdges } from './useSuggestedEdges';
+export { useNodeViolations } from './useNodeViolations';
+export { useEdgeViolations } from './useEdgeViolations';